Stop using valgrind with gcc 5+
ctest hard codes the use of the leak checker in valgrind, but this
produces leak reports from malloc in libstdc++-6 that are probably not
a problem, but not great to suppress unilaterally. Since we have
several other forms of leak checking (cppcheck, LeakSanitizer as part
of AddressSanitizer), we can do without valgrind in gcc 5.

Refactored the way we handle CMAKE_BUILD_TYPE so the behaviour of
"default to Debug build" is explicit and we don't have to do
double-negative tests, and repetitive checks whether CMAKE_BUILD_TYPE
is defined.

Refactored the inconsistent management of ensuring spaces in
opts_list, so that code is less sensitive to the existing order of
operations. Now every appended string prepends a space. Adopting the
opposite convention would also have bumped a number of different
lines.

Change-Id: I96c8c11ece9995265a8f1e80e41790721ba86f43