Support for specifying matrix in source repo
Add another entry point to the releng package that can be used to read
in a file with a list of configurations to build, and to convert that
into a format suitable for launching Jenkins jobs with Parameterized
Trigger plugin to build that list of configurations.
Also, add a helper function for getting the build url from Parameterized
Trigger back to the upstream job.

Does not (yet) select hosts for the configurations, but that can be
added without significantly influencing the Jenkins job configuration.

Change-Id: I25056f3a030f73ad5d89c21fc99ebafcfa3a9d8e