Restructure docs for inclusion in dev guide
Restructure the releng documentation such that it provides a single
entry point (index.rst), and provides an extensible interface to get the
list of files into an external CMake system.  Minimal fixes to get the
documentation to build without errors with sphinx.ext.autodoc.

This makes it possible to include the documentation into the dev guide
build from the main repository, and hopefully provides a stable enough
interface between the repositories so that the documentation here can
evolve easily without breaking backward compatibility for the
documentation build.

Change-Id: If30ec5c8fec62a1035caa52beed95ab79d64593c