Initial tests for some parts of releng
Add some unit tests for the contents of matrixbuild.py, and refactor the
code to support testing with full mocking of the environment and without
requirements on file system contents.

The tests require a separately installed backport of unittest.mock to
run on Python 2.7 (which is currently used on Jenkins), but this is
likely the most future-proof approach, since starting from Python 3.3,
this library is part of Python standard library.

Change-Id: I903ff90a9fdc816ad7a81f771b02a6d658ca14d1