Script to assign hosts to build configurations
Make the script that assembles the matrix for Parameterized Trigger also
assign a host for each configuration, based on the build options.

Supported labels are a pretty comprehensive reflection of the status
quo on the slaves, but not all combinations that seem reasonable work
just yet. We need multiple fixes on slaves and in the code, but that's
easier when releng repo can test such.

Additional TODOs:
- Better error handling if the host assignment fails.
- Support for using SIMD options for host selection.

Change-Id: Iad876d4fb478d76fc4c8511c09abef056e4942be