Allow specifying build options in build script
This removes the need to explicitly call a method in BuildEnvironment to
select the compiler.  Instead, build scripts can also specify a list of
build options, similar to the configuration matrixes.  This removes
multiple different ways for doing the same thing, and makes it simpler
to refactor handling of the the options.

Add some basic tests for the build script code.

Change-Id: I6da906a1232bdc8c22e796341db464adfc6d5816