Generalize passing build options to build script
Add context.opts that provides access to all build option values through
a single class, in a manner that it is not necessary to have a different
handler for each option to set a separate attribute in BuildParameters.
Similarly, it is no longer necessary to provide initial None assignment
for each of option, but they are instead populated from the list of
handlers.

Initial step for allowing build scripts to provide their own options
(for cases where the releng script does not need to care).

BuildParameters can be removed once we no longer need to support build
scripts that use it.

Change-Id: I4a323bc3c87583b0fe21bb8e34e672348c4755dd