Support declaring build options in build scripts
Make it possible to extend the set of build options from the build
script.  Once we have adapted the build scripts and can get rid of
BuildParameters backwards-compatibility, we can make releng only aware
of those options that actually influence the host placement or that need
to interact with the build slave configuration.

The main driver for this change is making it possible to extend the
verification matrixes to new types of configurations with changes only
in the source repo where both the configuration matrix and the build
script reside.  New types of options (e.g., with validated enum values
as a string value) can be added when necessary.  For now, just the basic
option types already available in options.py are supported.

Change-Id: I1edbbaee95edd25fa2b7aff32e3ed20077e04bfa