Support builds from tarballs
Make it possible to extract the source code and/or regression tests from
a tarball instead of checking out from git, and then using the standard
releng build scripts to do the actual build.

Make prefix option for regression tests packaging work as it should.

Also, centralize the handling of the package-info.log file used for this
purpose into releng: provide a helper method for writing this file, so
that the format and exact contents can be customized to what other parts
of releng expect.

Change-Id: Ic539f43255f587be6dc8ac59c8bbbe2fa76abf4d