Refactor refspec management
Clean up the management of information about the refspecs used to check
out projects, and about the state of the projects.  This enables future
encapsulation of different ways of getting the sources (git or tarball),
as well as some additional support for Jenkins workflow builds.  Both
will be done in child changes.

Change-Id: Iea07026febd14c85b91c61864a1d568f01de342f