Initial workflow for building release tarballs
This now covers most of what the existing Deploy_* jobs do, but with
some additional logic and a different kind of summary on the Jenkins
build page.

Remaining TODO items are identified in the script, but most critical
missing functionality only needs changes in the gromacs.py build script
in the source repo.  In the long run, testing added there may get
additional support from releng, but that is probably not necessary
initially.

Change-Id: Ibeae52917f59b7c7db5ce47262aa20422cfa445d