Do not use os.chdir for command execution dir
Make a separate object keep the working directory set with chdir(), and
pass this to each external command.  This allows better mocking of the
calls for future testing, and also makes it possible to clean the build
directory for out-of-source builds by deleting and recreating it,
without worrying about whether it is the current directory.

Also, resolve one TODO about a suboptimal default value.

Change-Id: I12e43805462f23cef2636299c71ea98ea893c445