Initial workflow support + docs updates
Add some basic support for creating workflow jobs using the releng
scripts, and update the documentation to describe the approach.

Also, add a workflow script for running a matrix job with configurations
specified in the source repository.  Compared to the current solution
based on Parameterized Trigger plugin, this has some advantages:
- It does not take up an extra executor slot while the matrix job is
  running, since there is only the flyweight executor for the workflow
  involved.
- The build status from the matrix build can be reliably propagated
  back.  This hopefully improves handling of aborted runs.
- The code can be easier to understand than the current complicated
  sequence of build steps that uses Parameterized Trigger.
However, because of JENKINS-32692, we cannot yet take this into use.

Change-Id: I06388996e158e395da6af389a8a42df24c105e87