Support writing out build matrix as JSON
This makes it easier to process the matrix in a workflow build.
Also, represent each configuration as a data structure instead of a
plain list of options, which clarifies the code overall.

Not used yet, but will be in the release workflow.  Split out to make
the main commit smaller.

Change-Id: I50989643c32d8f0dc6a87807039a4e95d0fbb147