Support for basic build parameter parsing
Add basic support for passing values from build parameters directly to
build scripts.  This is not ideal, as this creates some coupling between
the Jenkins job configuration and the build scripts, but this is mainly
for simplifying the setup for less common builds (for now, just the
packaging builds), and these do not have the same backwards
compatibility requirements as the per-patchset builds.

Change-Id: I9e8d887c99c889412dfb497cdcdc0dcffdf12a8a