Update clang static analyzer support
This is intended to be used on bs_nix1404, which now has clang 3.8.
It never had some other clang versions, so removed those. Added
new attribute clang-static-analyzer-3.8 so we can support
both analyzers during a transition period.

Used the new GROMACS standard-library selection functionality to
select libc++ for this build, which is easier to set up and maintain
than libstdc++.

Change-Id: I1f3ae9b678853be42e6c1b0052a463de0a827789