Make matrix launcher workflow work
With Gerrit Trigger 2.19.0, the workflow job should have all the
features necessary to fully replace the previous approach based on
Parameterized Trigger.  The workflow has two immediate advantages:
 - It does not take up an executor slot while the matrix build is
   building.
 - It should more reliably propagate back the status of the matrix
   build; the Parameterized Trigger version reports the build status as
   FAILURE even if the matrix build was actually aborted manually.

Additionally, with this approach, it is clear how, in principle, we
could also post additional details back to Gerrit about what exactly
failed in the matrix build.  But that will take a significant amount of
effort (in particular, to make the Python script aware of all the
possible ways the matrix build can become unstable, or alternatively get
that information otherwise to the parent workflow job).

Change-Id: I3aaffaf859bc17f0959de410ae8fca857a45bfec