Make crashing tests more visible in Jenkins
Produce a JUnit xml report from the CTest results, so that failures such
as crashes or gmx_fatal() show up in Jenkins also elsewhere than in the
console log.  Also initial adaptation to do the same for
AddressSanitizer results.

This could in principle also use XSLT, like it used to do, but in the
future it could be nice to also otherwise parse information about the
test results in the script.  Also, xsltproc can be an additional
maintenance burden that this approach avoids.

Change-Id: I595223e7aac179fd596fdf14755b8250e99a517c