Better support for changing compiler for coverage
Now that bs_mac has version-specific symlinks for gcov, use those to
make it possible to change the compiler used for the coverage build.

Also add a gcov-X.Y option that can be used to select the build slave,
although that is not (yet) necessary (the builds are still hard-coded
for bs_mac).

Change-Id: I8cc96a787afe307e18778d6ac4af22c1eb8cbcbc