On-demand cross-verification
Make it possible to trigger (some kinds of) cross-verification builds
using a '[JENKINS] Cross-verify NNNN' comment in Gerrit, and have the
results posted automatically back to both changes.  For now, this only
triggers the pre-submit matrix build, making it mostly suitable for
cross-verification between gromacs and regressiontests repos.  But with
some additional features, this could also replace the current, somewhat
difficult-to-use cross-verification features for releng changes.

Change-Id: I9fa236515385c712559af5da97b4469c59ab20e1