On-demand trigger for releng with release branches
Make it possible to trigger the release branch per-patchset builds on
request.  Triggering them for every patch set is inconvenient, since
they often block the build queue from builds that would actually be
relevant for testing the changes.  This moves the responsibility a bit
to the uploader and reviewers to trigger these before merging if any
impact could be present, but if this becomes a problem, we can consider
alternative mechanisms.

Change-Id: I420f7260c8488c2d64a9c5650f9bcea642ab9b9b