Full cross-verify support for releng changes
Make it possible to trigger all the builds affected by releng changes
using the [JENKINS] Cross-verify comment, and remove old mechanisms for
the same purpose that were more difficult to use.  Also add a quiet mode
to the request, which can be used to reduce spamming unrelated changes
when testing releng changes.

Change-Id: I2920dcbe45ddcffba89923dc4c1656c8db2e547a