Support for regressiontests reference generation
Add support to upload changes from Jenkins back to Gerrit.
This will be used initially by a job to update regressiontests
data, but perhaps also for an automatic uncrustifier, etc.
Adapt the on-demand build mechanism to support triggering this.

Migrated and extended various docs.

Added .gitignore to ignore .pyc files

Change-Id: Id03217c159372a07f3df154ac62d7e85ec64b6ea