Consolidate setup of clang on build slaves
It doesn't make great sense to try to use clang with C++11 on slaves
whose default compilers are too old. So move all of those to Ubuntu
slaves. However, the PPAs cannot install 3.4 alongside any later
version, so use two such slaves. And 3.4 can't build with sufficiently
recent libstdc++, and we want to build with an early version of clang
to help ensure portability, so choose one that doesn't have modern
gcc either.

Change-Id: I2cbf3fc6871046c3d158a36bcd882e351188293e