Restructure on-demand trigger comment parsing
Make the parser more structured by putting the logic in a class that
keeps state in member attributes, and splitting the logic into multiple
methods.  Change the accepted syntax a bit to ensure that the behavior
is predictable.  Make it possible to trigger any combination of builds
for cross-verification (e.g., only the cppcheck build).  Update
documentation to match.

Change-Id: I0efe614233cd2c61da09983991961f400d4e090b