Simplify checkouts in workflow builds
Since the workflow builds assume that every time a Python script is
invoked, CHECKOUT_PROJECT should be releng, structure the code such that
this always holds.  For the cases that want to use CHECKOUT_PROJECT like
it is in freestyle jobs (where it follows GERRIT_PROJECT and/or
project-specific default), introduce a separate variable that does not
interfere with other workflow code.

Change-Id: Ia5f6fba6f19660dc4af9e882195f08d9dbe6f814