Refactor passing information from Python to Groovy
Instead of using a named temporary file to pass output information from
Python code called from Groovy, use the status.json file that is already
created for passing back success/failure status and failure reason.
This possibly makes some of the Groovy code easier to follow when they
more explicitly use a "return value".

Document the structure of the returned objects in Groovy, so that it is
not necessary to decipher them from surrounding code or look into the
Python code.

Minor fixes for issues noticed:
- Consistently check the assigned matrix hosts (also for on-demand
  builds).
- Always initialize NODE_NAME to a string value in __main__.py, as
  otherwise this test mode does not work.

Change-Id: Iab5a0d38130306197aac2955d9493305fca8b012