Centralize some Jenkins env.var. management
Centralize the logic of reading the WORKSPACE and NODE_NAME environment
variables from Jenkins into integration.py.  This moves a bit more logic
of interacting with Jenkins into integration.py, making it easier to see
all of it in a single place. It also makes it clearer how to access the
name of the build slave if it is needed for some logic.

Change-Id: I9a3d27fc4be3ca99c9fb09941dcdc6d93d2cfc45