Add clang-4 to a GPU slave
Creates the possibility to test the new experimental CUDA+clang
support. Note that recent clang packages support OpenMP, and releng
already knows how to set up the linking to libomp.so.

clang-3.3 is no longer present on that slave (or used anywhere).

Disabled clang-4 on the other CUDA slave so that we get the job
sent to the less heavily loaded slave.

Change-Id: Idbe2fd9cd30cfe982dadd5c319f102b2aa77eb99