Updated organization for linking libstdc++
gcc builds now always link to the libraries installed alongside
them. This also ensures that e.g. sanitizer builds link suitably.

On Linux and mac, clang and icc builds now link to libraries from a
specified gcc installed on that host.

Made find_executable return the real path, since we normally use
it in a way that suits that behaviour.

Resolves some TODO items.

Note that if/when we want to test specific compiler+library+abi
setups, we will do that with specifiers in testing matrices, not
via these defaults.

Change-Id: If00fb78a6064f01de863e192aa8160cd6faa39da