Improve version handling for on-demand regtest package
Use the source code version number instead of "test" as the version
number for the on-demand regressiontests package testing.  For most uses,
this doesn't matter, but it makes things appear a bit more consistent,
and it's easy enough extension to existing code.

For testing, make it possible to trigger only the regressiontests
package for releng changes.

Change-Id: Id6318abafee6a88125777aad9a05f6d9dea503b3