Move TSAN jobs to non-VM slave
These were taking about twice as long as other verification stages,
probably because bs_nix1404 is on a VM, and on a VM host whose cores
are quite slow. By moving the TSAN support to a non-VM slave,
the TSAN verify is now no longer rate limiting, and we can
also later test TSAN with SIMD if we prefer.

On bs_mic, the system gcc is 4.4 and the ld also can't link with newer
GCC, so refactored the way we set up the default shell, so that this
slave can use the devtoolset. This means we can have a build
environment that can function well, and can be customized to suit the
needs of particular Jenkins slaves. Most slaves just do what they
always did.

Shifted the responsibility for constructing a command that might
modify the environment and then dump it to the callers of import_env.

Change-Id: I22eb74e0f9ed6f5bee7599ceeace9361436861c0