Add icc 17 and 18 to bs_mic
Updated the script to gracefully handle new versions of icc if they
conform to the current pattern, so we don't have to keep updating
releng for this.

Note that we do not intend to specify minor versions of compilers in
matrices, so Intel compilers should be referred to by their year, not
year and patch level. If we subseqeuently install a later patch level
for some year, we plan to test in Jenkins with the later one only.

Change-Id: Ic1f9bd6996b00ed65ddce936d9ade54e4167163f