Make it possible to test that a non-GPU build works
Arranged the slave configuration descriptions such that the only place
that CUDA 7.0 is available is a slave that has no GPU, so that we can
have a configuration that tests that a CUDA build passes tests even
when there is no GPU detected. It would be preferable to separate
matrix descriptions for build-time requirements (e.g. a CUDA version)
from execution-time requirements (absence/presence of a device that
supports running CUDA code).

Refs #2161

Change-Id: Ia4a352b19b08add8867de47dc423f39c1117419b