Refactor ProjectsManager
Keep only a single dictionary that records all information about the
projects. This makes the state inside the class (and for ProjectInfo)
more explicit, and makes it easier to manage more complex state
internally. Improve related tests to assert more (even more could be
added, but not done for now) and to not rely on a method that is no
longer needed for anything else.

Change-Id: Ia564514aff2725cac06432f1699c143cacc212f0