Improve releng unit tests
Add control over what is returned by various command invocations (in
particular, git rev-list, git ls-remote, and ssh gerrit query) so that
they can return more realistic information.  Implement a test helper
that keeps information about the desired (mock) commit for each project,
and use that to serve the queries as well as for constructing expected
results for tests.

Remove unnecessary uses of CHECKOUT_PROJECT etc. now that it is not
mandatory.  Simplify use of 'workspace' parameter to TestHelper, only
pass it in tests where actual absolute workspace paths are used, and
always pass the same string.  There can be room for more improvement
here, but the change is large enough as it is.

Change-Id: I7798831789110f67860ab935c7f4551717bcc03a