Rearrange stdlib handling
For GROMACS 2020, we will build with a specified C++ standard library
in a different way. This requires reorganization so that build options
have all been processed before we prepare the way the standard library
will be set up.

Maintenance branches continue to work in the old way, and comments
have been left to clarify which code can be removed when it is no
longer required.

Change-Id: I0c947475b429c6cca86a7871efe2799a9781c4a1